Move edit project management form into modules directory
