+++ src/apps/investments/client/projects/team/state.js
@@ -1,3 +1,1 @@
 export const TASK_EDIT_PROJECT_TEAM_MEMBERS = 'TASK_EDIT_PROJECT_TEAM_MEMBERS'
-export const TASK_SAVE_INVESTMENT_PROJECT_MANAGERS =
-  'TASK_SAVE_INVESTMENT_PROJECT_MANAGERS'
