+++ src/client/modules/Investments/Projects/state.js
@@ -29,6 +29,8 @@ export const TASK_UPDATE_INVESTMENT_PROJECT_STAGE =
   'TASK_UPDATE_INVESTMENT_PROJECT_STAGE'
 export const TASK_SAVE_CLIENT_RELATIONSHIP_MANAGER =
   'TASK_SAVE_CLIENT_RELATIONSHIP_MANAGER'
+export const TASK_SAVE_INVESTMENT_PROJECT_MANAGERS =
+  'TASK_SAVE_INVESTMENT_PROJECT_MANAGERS'
 
 export const TASK_GET_NON_FDI_PROJECTS_LIST = 'TASK_GET_NON_FDI_PROJECTS_LIST'
 export const TASK_UPDATE_ASSOCIATED_PROJECT = 'TASK_UPDATE_ASSOCIATED_PROJECT'
