+++ src/client/routes.js
@@ -38,7 +38,7 @@ import ProjectEvaluation from './modules/Investments/Projects/ProjectEvaluation'
 import ProjectInteractions from './modules/Investments/Projects/ProjectInteractions'
 import ProjectPropositions from './modules/Investments/Projects/ProjectPropositions'
 import EditClientRelationshipManagement from './modules/Investments/Projects/EditClientRelationshipManagement'
-import EditProjectManagement from '../apps/investments/client/projects/team/EditProjectManagement'
+import EditProjectManagement from './modules/Investments/Projects/EditProjectManagement'
 
 const routes = {
   companies: [
