+++ src/client/tasks.js
@@ -103,10 +103,7 @@ import {
 } from '../apps/companies/apps/company-overview/overview-table-cards/state'
 import * as overviewInvestmentProjectTasks from '../apps/companies/apps/company-overview/overview-table-cards/tasks'
 
-import {
-  TASK_EDIT_PROJECT_TEAM_MEMBERS,
-  TASK_SAVE_INVESTMENT_PROJECT_MANAGERS,
-} from '../apps/investments/client/projects/team/state'
+import { TASK_EDIT_PROJECT_TEAM_MEMBERS } from '../apps/investments/client/projects/team/state'
 import * as editInvestmentProjectTeamTasks from '../apps/investments/client/projects/team/tasks'
 
 import {
@@ -302,6 +299,7 @@ import {
   TASK_GET_UK_COMPANIES,
   TASK_UPDATE_RECIPIENT_COMPANY,
   TASK_SAVE_CLIENT_RELATIONSHIP_MANAGER,
+  TASK_SAVE_INVESTMENT_PROJECT_MANAGERS,
 } from './modules/Investments/Projects/state'
 import {
   getNonFdiProjects,
